requirements.txt
discord.py
shelves.py
